refactor: domain 내부 클래스 구분
