+++ src/main/java/domain/lotto/vo/Prize.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.lotto.vo;
 
 import java.util.Arrays;
