+++ src/test/java/domain/machine/LottoMachineTest.java
@@ -1,7 +1,10 @@
-package domain;
+package domain.machine;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.machine.LottoMachine;
+import domain.lotto.vo.LottoResult;
+import domain.lotto.vo.WinNumbers;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
